third_party/WebKit/LayoutTests/inspector-protocol/layers/get-layers-expected.txt


=== Get the Document ===

PASS

=== Enable the LayerTree agent ===


[
    {
        "layerId": "string",
        "offsetX": 0,
        "offsetY": 0,
        "width": "number",
        "height": "number",
        "paintCount": "number",
        "nodeId": "number",
        "parentLayerId": "string"
    },
    {
        "layerId": "string",
        "offsetX": 25,
        "offsetY": 25,
        "width": 50,
        "height": 50,
        "paintCount": "number",
        "nodeId": "number",
        "parentLayerId": "string"
    },
    {
        "layerId": "string",
        "offsetX": 25,
        "offsetY": 25,
        "width": 50,
        "height": 50,
        "paintCount": "number",
        "nodeId": "number",
        "parentLayerId": "string"
    },
    {
        "layerId": "string",
        "offsetX": 200,
        "offsetY": 0,
        "width": 100,
        "height": 100,
        "paintCount": "number",
        "nodeId": "number",
        "parentLayerId": "string"
    },
    {
        "layerId": "string",
        "offsetX": 25,
        "offsetY": 25,
        "width": 50,
        "height": 50,
        "paintCount": "number",
        "nodeId": "number",
        "parentLayerId": "string"
    },
    {
        "layerId": "string",
        "offsetX": 0,
        "offsetY": 0,
        "width": "number",
        "height": "number",
        "paintCount": "number",
        "nodeId": "number",
        "parentLayerId": "string"
    }
]

=== Message the page to add a new composited layer ===


[
    {
        "layerId": "string",
        "offsetX": 0,
        "offsetY": 0,
        "width": "number",
        "height": "number",
        "paintCount": "number",
        "nodeId": "number",
        "parentLayerId": "string"
    },
    {
        "layerId": "string",
        "offsetX": 25,
        "offsetY": 25,
        "width": 50,
        "height": 50,
        "paintCount": "number",
        "nodeId": "number",
        "parentLayerId": "string"
    },
    {
        "layerId": "string",
        "offsetX": 25,
        "offsetY": 25,
        "width": 50,
        "height": 50,
        "paintCount": "number",
        "nodeId": "number",
        "parentLayerId": "string"
    },
    {
        "layerId": "string",
        "offsetX": 200,
        "offsetY": 0,
        "width": 100,
        "height": 100,
        "paintCount": "number",
        "nodeId": "number",
        "parentLayerId": "string"
    },
    {
        "layerId": "string",
        "offsetX": 25,
        "offsetY": 25,
        "width": 50,
        "height": 50,
        "paintCount": "number",
        "nodeId": "number",
        "parentLayerId": "string"
    },
    {
        "layerId": "string",
        "offsetX": 0,
        "offsetY": 0,
        "width": "number",
        "height": "number",
        "paintCount": "number",
        "nodeId": "number",
        "parentLayerId": "string"
    },
    {
        "layerId": "string",
        "offsetX": 25,
        "offsetY": 25,
        "width": 50,
        "height": 50,
        "paintCount": "number",
        "nodeId": "number",
        "parentLayerId": "string"
    }
]

=== Get attributes for the newly inserted node ===

PASS
DONE!
